doctor-analyzer/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
boto3>=1.33.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.0
httpx>=0.25.0
